set up rescheduling-prompt endpoint
