+++ notification-svc/routes/rescheduling-prompt.ts
@@ -1,35 +1,13 @@
-import express, { Response, Request } from 'express'
-import sendEmail from '../models/sendEmail'
-import { reschedulingEventEmitter } from '../utils/customEventEmitters'
+import express from 'express'
+import reschedulingController from '../controllers/reschedulingController'
 
 const reschedulingPromptRoutes = express.Router()
 
 // endpoint for rescheduling-svc to interact with
-reschedulingPromptRoutes.post('/rescheduling-prompt', (req: Request, res: Response) => {
-  const { patient, appointment } = req.body
-  let isPending = true
-  sendEmail(`Dear ${patient}, would you like to take an earlier appointment ${appointment}.`)
-  isPending && reschedulingEventEmitter.once('prompt-handled', message => {
-    res.status(200).send(message);
-    isPending = false
-  })
-  setTimeout(() => {
-    isPending && (reschedulingEventEmitter.emit = () => false)
-    isPending && res.status(200).send('expired')
-    isPending = false
-  }, 1000)
-})
+reschedulingPromptRoutes.post('/rescheduling-prompt', reschedulingController.prompt)
 
-reschedulingPromptRoutes.post('/rescheduling-confirm', (req: Request, res: Response) => {
-  console.log('rescheduling confirmed')
-  reschedulingEventEmitter.emit('prompt-handled', 'confirmed')
-  res.status(200).send('Your rescheduling confirmed')
-})
+reschedulingPromptRoutes.post('/rescheduling-confirm', reschedulingController.confirm)
 
-reschedulingPromptRoutes.post('/rescheduling-reject', (req: Request, res: Response) => {
-  console.log('rescheduling rejected')
-  reschedulingEventEmitter.emit('prompt-handled', 'rejected')
-  res.status(200).send('Your rescheduling rejected')
-})
+reschedulingPromptRoutes.post('/rescheduling-reject', reschedulingController.reject)
 
 export default reschedulingPromptRoutes
